changed to long long for windows

+++ inc/hoymiles/dtu.h
@@ -18,7 +18,7 @@ class Dtu {
 
 	void populateMicroinverters();
 
-	std::pair<bool, Microinverter*> getMicroinverterBySerialNumber(long serialNumber);
+	std::pair<bool, Microinverter*> getMicroinverterBySerialNumber(long long serialNumber);
 
       public:
 	Dtu(const char *ip_address, int port);
@@ -27,11 +27,11 @@ class Dtu {
 
 	// void updateMicroinverters();
 
-	void updateMicroinverters(std::vector<std::string> &parametersToGet, bool allParameters, std::vector<long> &microinvertersToGet);
+	void updateMicroinverters(std::vector<std::string> &parametersToGet, bool allParameters, std::vector<long long> &microinvertersToGet);
 
 	// void printMicroinverters();
 
-	void printMicroinverters(std::vector<std::string> &parametersToGet, bool allParameters, std::vector<long> &microinvertersToGet);
+	void printMicroinverters(std::vector<std::string> &parametersToGet, bool allParameters, std::vector<long long> &microinvertersToGet);
 
 	~Dtu();
 };
